+++ app/Hutch.Relay/Commands/AddUser.cs
@@ -1,28 +1,101 @@
 using System.CommandLine;
-using Hutch.Relay.Commands.Helpers;
+using System.CommandLine.Invocation;
+using Hutch.Relay.Data.Entities;
+using Hutch.Relay.Helpers;
+using Hutch.Relay.Services.Contracts;
+using Hutch.Relay.Startup.Cli.Core;
+using Microsoft.AspNetCore.Identity;
+using Spectre.Console;
 
 namespace Hutch.Relay.Commands;
 
 internal class AddUser : Command
 {
+  public static readonly Argument<string> Username =
+    new("username")
+    {
+      Description = "The new user name."
+    };
+
   public AddUser(string name)
     : base(name, "Add a new User.")
   {
-    var argUserName = new Argument<string>("username", "The new user name.");
-    Add(argUserName);
-    
-    this.SetHandler(
-      async (
-        scopeFactory,
-        username) =>
+    Arguments.Add(Username);
+
+    Action = CliApplication.AsyncAction<AddUserAction>();
+  }
+}
+
+internal class AddUserAction(
+  [FromKeyedServices("stdout")] IAnsiConsole stdout,
+  [FromKeyedServices("stderr")] IAnsiConsole stderr,
+  UserManager<RelayUser> users,
+  ISubNodeService subNodes)
+  : AsynchronousCommandLineAction
+{
+
+  public override async Task<int> InvokeAsync(ParseResult parseResult, CancellationToken cancellationToken = default)
+  {
+    var username = parseResult.GetValue(AddUser.Username);
+
+    if (string.IsNullOrWhiteSpace(username))
+    {
+      stderr.MarkupLine("[red]:warning: Username cannot be empty.[/]");
+      return 1;
+    }
+
+    if (await users.FindByNameAsync(username) is not null)
+    {
+      stderr.MarkupLine($"[red]:warning: User {username} already exists.[/]");
+      return 1;
+    }
+    {
+      stderr.Write(new Rule("[green]Add User[/]")
       {
-        using var scope = scopeFactory.CreateScope();
-        
-        var runner = scope.ServiceProvider.GetRequiredService<Runners.AddUser>();
-        
-        await runner.Run(username);
-      },
-      Bind.FromServiceProvider<IServiceScopeFactory>(),
-      argUserName);
+        Justification = Justify.Left
+      });
+
+      var user = new RelayUser()
+      {
+        UserName = username
+      };
+
+      var password = GeneratePassword.GenerateUniquePassword(16);
+
+      var result = await users.CreateAsync(user, password);
+
+
+      if (!result.Succeeded)
+      {
+        stderr.MarkupLine($"[red]:warning: User creation failed with errors for {username}.[/]");
+
+        var errorRows = result.Errors
+          .Select(e => new Text(e.Description, new(Color.Red, Color.Black)))
+          .ToList();
+
+        stderr.Write(new Rule("[red]Errors[/]")
+        {
+          Justification = Justify.Left
+        });
+
+        stderr.Write(new Rows(errorRows));
+
+        return 1;
+      }
+
+      var subNode = await subNodes.Create(user);
+
+      var table = new Table { Border = TableBorder.Rounded };
+
+      table.AddColumn("[blue]Username[/]");
+      table.AddColumn("[blue]Password[/]");
+      table.AddColumn("[blue]Collection ID[/]");
+
+      table.AddRow(username, password, subNode.Id.ToString());
+
+      stdout.Write(table);
+
+      return 0;
+    }
   }
 }
