+++ app/Hutch.Relay/Commands/AddUserSubNode.cs
@@ -1,34 +1,34 @@
-using System.CommandLine;
-using Hutch.Relay.Commands.Helpers;
+// using System.CommandLine;
+// using Hutch.Relay.Commands.Helpers;
 
-namespace Hutch.Relay.Commands;
+// namespace Hutch.Relay.Commands;
 
-internal class AddUserSubNode : Command
-{
-  public AddUserSubNode(string name)
-    : base(name, "Add a new Sub Node for a User.")
-  {
-    var argUserName = new Argument<string>("username", "The User to create a Sub Node for.");
-    Add(argUserName);
+// internal class AddUserSubNode : Command
+// {
+//   public AddUserSubNode(string name)
+//     : base(name, "Add a new Sub Node for a User.")
+//   {
+//     var argUserName = new Argument<string>("username", "The User to create a Sub Node for.");
+//     Add(argUserName);
 
-    var optAutoConfirm = new Option<bool>("--yes", "Automatically say yes to the final confirmation check.");
-    optAutoConfirm.AddAlias("-y");
-    Add(optAutoConfirm);
+//     var optAutoConfirm = new Option<bool>("--yes", "Automatically say yes to the final confirmation check.");
+//     optAutoConfirm.AddAlias("-y");
+//     Add(optAutoConfirm);
 
-    this.SetHandler(
-      async (
-        scopeFactory,
-        username,
-        autoConfirm) =>
-      {
-        using var scope = scopeFactory.CreateScope();
+//     this.SetHandler(
+//       async (
+//         scopeFactory,
+//         username,
+//         autoConfirm) =>
+//       {
+//         using var scope = scopeFactory.CreateScope();
 
-        var runner = scope.ServiceProvider.GetRequiredService<Runners.AddUserSubNode>();
+//         var runner = scope.ServiceProvider.GetRequiredService<Runners.AddUserSubNode>();
 
-        await runner.Run(username, autoConfirm);
-      },
-      Bind.FromServiceProvider<IServiceScopeFactory>(),
-      argUserName,
-      optAutoConfirm);
-  }
-}
+//         await runner.Run(username, autoConfirm);
+//       },
+//       Bind.FromServiceProvider<IServiceScopeFactory>(),
+//       argUserName,
+//       optAutoConfirm);
+//   }
+// }
