+++ app/Hutch.Relay/Commands/DatabaseUpdate.cs
@@ -1,21 +1,41 @@
 using System.CommandLine;
-using Hutch.Relay.Commands.Helpers;
+using System.CommandLine.Invocation;
+using Hutch.Relay.Services;
+using Hutch.Relay.Startup.Cli.Core;
+using Spectre.Console;
 
 namespace Hutch.Relay.Commands;
 
 internal class DatabaseUpdate : Command
 {
-  public DatabaseUpdate(string name) : base(name, "Update the database to the latest migration in this build.")
+  public DatabaseUpdate(string name) : base(name,
+    "Update the database to the latest migration in this build.")
   {
-    this.SetHandler(
-      async (scopeFactory) =>
-      {
-        using var scope = scopeFactory.CreateScope();
+    Action = CliApplication.AsyncAction<DatabaseUpdateAction>();
+  }
+}
 
-        var runner = scope.ServiceProvider.GetRequiredService<Runners.DatabaseUpdate>();
+internal class DatabaseUpdateAction(
+  DbManagementService dbManager,
+  [FromKeyedServices("stdout")] IAnsiConsole stdout,
+  [FromKeyedServices("stderr")] IAnsiConsole stderr)
+  : AsynchronousCommandLineAction
+{
+  public override async Task<int> InvokeAsync(ParseResult parseResult, CancellationToken cancellationToken = default)
+  {
+    var rule = new Rule("[green]Update Database[/]")
+    {
+      Justification = Justify.Left
+    };
+    stderr.Write(rule);
 
-        await runner.Run();
-      },
-      Bind.FromServiceProvider<IServiceScopeFactory>());
+    await stderr.Status()
+      .StartAsync("Updating Database to the latest migration...",
+        async context => { await dbManager.UpdateDatabase(); });
+
+    // If anything goes wrong, the manager will throw
+    stdout.MarkupLine("[blue]The Relay Database is up to date![/]");
+
+    return 0;
   }
 }
