+++ app/Hutch.Relay/Commands/ListUserSubNodes.cs
@@ -1,28 +1,28 @@
-using System.CommandLine;
-using Hutch.Relay.Commands.Helpers;
+// using System.CommandLine;
+// using Hutch.Relay.Commands.Helpers;
 
-namespace Hutch.Relay.Commands;
+// namespace Hutch.Relay.Commands;
 
-internal class ListUserSubNodes : Command
-{
-  public ListUserSubNodes(string name)
-    : base(name, "List all Sub Nodes for a User.")
-  {
-    var argUserName = new Argument<string>("username", "The User to list Sub Nodes for.");
-    Add(argUserName);
+// internal class ListUserSubNodes : Command
+// {
+//   public ListUserSubNodes(string name)
+//     : base(name, "List all Sub Nodes for a User.")
+//   {
+//     var argUserName = new Argument<string>("username", "The User to list Sub Nodes for.");
+//     Add(argUserName);
 
-    this.SetHandler(
-      async (
-        scopeFactory,
-        username) =>
-      {
-        using var scope = scopeFactory.CreateScope();
+//     this.SetHandler(
+//       async (
+//         scopeFactory,
+//         username) =>
+//       {
+//         using var scope = scopeFactory.CreateScope();
 
-        var runner = scope.ServiceProvider.GetRequiredService<Runners.ListUserSubNodes>();
+//         var runner = scope.ServiceProvider.GetRequiredService<Runners.ListUserSubNodes>();
 
-        await runner.Run(username);
-      },
-      Bind.FromServiceProvider<IServiceScopeFactory>(),
-      argUserName);
-  }
-}
+//         await runner.Run(username);
+//       },
+//       Bind.FromServiceProvider<IServiceScopeFactory>(),
+//       argUserName);
+//   }
+// }
