+++ app/Hutch.Relay/Commands/ListUsers.cs
@@ -1,23 +1,23 @@
-using System.CommandLine;
-using Hutch.Relay.Commands.Helpers;
+// using System.CommandLine;
+// using Hutch.Relay.Commands.Helpers;
 
-namespace Hutch.Relay.Commands;
+// namespace Hutch.Relay.Commands;
 
-internal class ListUsers : Command
-{
-  public ListUsers(string name)
-    : base(name, "List Users.")
-  {
-    this.SetHandler(
-      async (
-        scopeFactory) =>
-      {
-        using var scope = scopeFactory.CreateScope();
+// internal class ListUsers : Command
+// {
+//   public ListUsers(string name)
+//     : base(name, "List Users.")
+//   {
+//     this.SetHandler(
+//       async (
+//         scopeFactory) =>
+//       {
+//         using var scope = scopeFactory.CreateScope();
         
-        var runner = scope.ServiceProvider.GetRequiredService<Runners.ListUsers>();
+//         var runner = scope.ServiceProvider.GetRequiredService<Runners.ListUsers>();
         
-        await runner.Run();
-      },
-      Bind.FromServiceProvider<IServiceScopeFactory>());
-  }
-}
+//         await runner.Run();
+//       },
+//       Bind.FromServiceProvider<IServiceScopeFactory>());
+//   }
+// }
