+++ app/Hutch.Relay/Commands/ResetUserPassword.cs
@@ -1,28 +1,28 @@
-using System.CommandLine;
-using Hutch.Relay.Commands.Helpers;
+// using System.CommandLine;
+// using Hutch.Relay.Commands.Helpers;
 
-namespace Hutch.Relay.Commands;
+// namespace Hutch.Relay.Commands;
 
-internal class ResetUserPassword : Command
-{
-  public ResetUserPassword(string name)
-    : base(name, "Reset a User's password.")
-  {
-    var argUserName = new Argument<string>("username", "The user to create a new password for.");
-    Add(argUserName);
+// internal class ResetUserPassword : Command
+// {
+//   public ResetUserPassword(string name)
+//     : base(name, "Reset a User's password.")
+//   {
+//     var argUserName = new Argument<string>("username", "The user to create a new password for.");
+//     Add(argUserName);
 
-    this.SetHandler(
-      async (
-        scopeFactory,
-        username) =>
-      {
-        using var scope = scopeFactory.CreateScope();
+//     this.SetHandler(
+//       async (
+//         scopeFactory,
+//         username) =>
+//       {
+//         using var scope = scopeFactory.CreateScope();
         
-        var runner = scope.ServiceProvider.GetRequiredService<Runners.ResetUserPassword>();
+//         var runner = scope.ServiceProvider.GetRequiredService<Runners.ResetUserPassword>();
         
-        await runner.Run(username);
-      },
-      Bind.FromServiceProvider<IServiceScopeFactory>(),
-      argUserName);
-  }
-}
+//         await runner.Run(username);
+//       },
+//       Bind.FromServiceProvider<IServiceScopeFactory>(),
+//       argUserName);
+//   }
+// }
